Crate minigrep

minigrep

This is my minigrep tool made only for learning rust. It is aimed to provided similar functionality with ripgrep. But for now it is still work in progress. Run slow, buggy and refactor-required. There are currently two crate here. First is argtool, for parse sys arg, used by minigrep. The other is grep, the main program for minigrep.

grep

Installation

# cd to this repo
cargo build
# This is binary location
./target/release/minigrep

Current usage

Simple usage:

Basic pattern

./minigrep "YourPattern" "YourDir/*"

Print all fn in codes.

./target/release/minigrep "fn" "minigrep/*" 

Print help

./target/release/minigrep -h
./target/release/minigrep --help

More usage:

Print with line number and the following two line.

./target/release/minigrep "fn" "minigrep/*" -n -B 2

Multiple match: struct, trait, and enum

./target/release/minigrep "trait" "minigrep/*" -n  -M "struct" -M "enum" -B 3 -A 1

Extract match pattern, here the pattern is function name 🪲: There are some wrong behavior in extract, need capture group. If parenthesis is added, that is "fn [A-Za-z0-9_]+\(" Some of matched line will be empty.

./target/release/minigrep "fn" "minigrep/*" -n -E "fn [A-Za-z0-9_]+"

Replace -> with ==> 🪲: if replacer is --> there will be error. - need to be escaped. We need "\->" for -R only, there is no need to specify it twice.

./target/release/minigrep "\->" "minigrep/*" -n -R "\->" -r "==>"

argtool

go to github page generated by rust docs

Modules

argparse

grep

utils

Functions

main_loop

main loop for grep a file